requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.11.5
cryptography==2.4.2
Django==2.1.5
idna==2.8
paramiko==2.4.2
Pillow==5.4.1
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.7.3
pycryptodomex==3.7.2
PyMySQL==0.9.3
PyNaCl==1.3.0
python-alipay-sdk==1.9.1
pytz==2018.9
six==1.12.0
